ocr_api/requirements.txt
flask==2.3.3
werkzeug==2.3.7
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
sentencepiece==0.1.99
python-Levenshtein==0.21.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
pillow==10.0.0
paddlepaddle==2.5.2  # Version mise à jour
paddleocr==2.6.1.3
pymupdf<1.21.0
python-doctr==0.6.0
scikit-learn==1.3.0
opencv-python<=4.6.0.66
huggingface-hub==0.16.4
gunicorn==21.2.0
pdf2image==1.16.3
fitz==0.0.1.dev2